# lib/api-zod/src/generated/types/child.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface Child {
  id: string;
  name: string;
  dob: string;
  startDate: string;
  createdAt: string;
  updatedAt: string;
  status?: string;
  archivedAt?: string | null;
  baselineStep?: number | null;
  isDemo?: boolean | null;
}