# lib/api-zod/src/generated/types/rating.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { HistoryEntry } from "./historyEntry";

export interface Rating {
  status: string;
  updatedAt: string;
  history?: HistoryEntry[];
}